Move free representation to its own module.
